Avoid a NullPointerException if file listing fails.
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/classlib/branches/java6@765253 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/luni/src/test/api/common/org/apache/harmony/luni/tests/java/io/FileTest.java
@@ -58,15 +58,17 @@ public class FileTest extends TestCase {
         // Delete all old temporary files
         File tempDir = new File(System.getProperty("java.io.tmpdir"));
         String[] files = tempDir.list();
-        for (int i = 0; i < files.length; i++) {
-            File f = new File(tempDir, files[i]);
-            if (f.isDirectory()) {
-                if (files[i].startsWith("hyts_resources")) {
-                    deleteTempFolder(f);
+        if (files != null) {
+            for (int i = 0; i < files.length; i++) {
+                File f = new File(tempDir, files[i]);
+                if (f.isDirectory()) {
+                    if (files[i].startsWith("hyts_resources")) {
+                        deleteTempFolder(f);
+                    }
                 }
-            }
-            if (files[i].startsWith("hyts_") || files[i].startsWith("hyjar_")) {
-                new File(tempDir, files[i]).delete();
+                if (files[i].startsWith("hyts_") || files[i].startsWith("hyjar_")) {
+                    new File(tempDir, files[i]).delete();
+                }
             }
         }
     }
